Make sure D'Arcy's elf/configure changes get into -stable:
-------
From: "D'Arcy" "J.M." Cain <darcy@druid.net>

I didn't see any further discussion so here is, I hope, a clean fix to
configure.in to determine if a system is ELF or not.  Note that some
of the tests earlier may be redundant but I took the safest route.