Fix DROP OWNED BY to correctly consider the implicitly-deleted objects list for each object to be deleted, instead of the previous hack that just skipped INTERNAL dependencies, which didn't really work.  Per report from Tom Lane.
To do this, introduce a new performMultipleDeletions entry point in
dependency.c to delete multiple objects at once.  The dependency code then has
the responsability of tracking INTERNAL and AUTO dependencies as needed.

Along the way, change ObjectAddresses so that we can allocate an ObjectAddress
list from outside dependency.c and not have to export the internal
representation.