Suppress useless searches for unused line pointers in PageAddItem.  To do this, add a 16-bit "flags" field to page headers by stealing some bits from pd_tli.  We use one flag bit as a hint to indicate whether there are any unused line pointers; the remaining 15 are available for future use.
This is a cut-down form of an idea proposed by Hiroki Kataoka in July 2005.
At the time it was rejected because the original patch increased the size of
page headers and it wasn't clear that the benefit outweighed the distributed
cost.  The flag-bit approach gets most of the benefit without requiring an
increase in the page header size.

Heikki Linnakangas and Tom Lane